# pkg/trace: use msgp.Marshaler instead of msgp.Decodable of msgpack deserialization