--- a/pkg/trace/api/api_test.go
+++ b/pkg/trace/api/api_test.go
@@ -455,7 +455,7 @@ func TestDecodeV05(t *testing.T) {
 	assert.NoError(err)
 	req, err := http.NewRequest("POST", "/v0.5/traces", bytes.NewReader(b))
 	assert.NoError(err)
-	traces, err := decodeTraces(v05, req, nil)
+	traces, err := decodeTraces(v05, req)
 	assert.NoError(err)
 	assert.EqualValues(traces, pb.Traces{
 		{