--- a/pkg/trace/pb/decoder.go
+++ b/pkg/trace/pb/decoder.go
@@ -228,3 +228,164 @@ func parseInt32(dc *msgp.Reader) (int32, error) {
 		return 0, msgp.TypeError{Encoded: t, Method: msgp.IntType}
 	}
 }
+
+// parseStringBytes reads the next type in the msgpack payload and
+// converts the BinType or the StrType in a valid string.
+func parseStringBytes(bts []byte) (string, []byte, error) {
+	// read the generic representation type without decoding
+	t := msgp.NextType(bts)
+
+	var err error
+	var i []byte
+	switch t {
+	case msgp.BinType:
+		i, bts, err = msgp.ReadBytesZC(bts)
+	case msgp.StrType:
+		i, bts, err = msgp.ReadStringZC(bts)
+	default:
+		return "", bts, msgp.TypeError{Encoded: t, Method: msgp.StrType}
+	}
+	if err != nil {
+		return "", bts, err
+	}
+	if utf8.Valid(i) {
+		return string(i), bts, nil
+	}
+	return repairUTF8(msgp.UnsafeString(i)), bts, nil
+}
+
+// parseFloat64Bytes parses a float64 even if the sent value is an int64 or an uint64;
+// this is required because the encoding library could remove bytes from the encoded
+// payload to reduce the size, if they're not needed.
+func parseFloat64Bytes(bts []byte) (float64, []byte, error) {
+	// read the generic representation type without decoding
+	t := msgp.NextType(bts)
+
+	var err error
+	switch t {
+	case msgp.IntType:
+		var i int64
+		i, bts, err = msgp.ReadInt64Bytes(bts)
+		if err != nil {
+			return 0, bts, err
+		}
+
+		return float64(i), bts, nil
+	case msgp.UintType:
+		var i uint64
+		i, bts, err = msgp.ReadUint64Bytes(bts)
+		if err != nil {
+			return 0, bts, err
+		}
+
+		return float64(i), bts, nil
+	case msgp.Float64Type:
+		var f float64
+		f, bts, err = msgp.ReadFloat64Bytes(bts)
+		if err != nil {
+			return 0, bts, err
+		}
+
+		return f, bts, nil
+	default:
+		return 0, bts, msgp.TypeError{Encoded: t, Method: msgp.Float64Type}
+	}
+}
+
+// parseInt64Bytes parses an int64 even if the sent value is an uint64;
+// this is required because the encoding library could remove bytes from the encoded
+// payload to reduce the size, if they're not needed.
+func parseInt64Bytes(bts []byte) (int64, []byte, error) {
+	// read the generic representation type without decoding
+	t := msgp.NextType(bts)
+
+	var i int64
+	var u uint64
+	var err error
+	switch t {
+	case msgp.IntType:
+		i, bts, err = msgp.ReadInt64Bytes(bts)
+		if err != nil {
+			return 0, bts, err
+		}
+		return i, bts, nil
+	case msgp.UintType:
+		u, bts, err = msgp.ReadUint64Bytes(bts)
+		if err != nil {
+			return 0, bts, err
+		}
+
+		// force-cast
+		i, ok := castInt64(u)
+		if !ok {
+			return 0, bts, errors.New("found uint64, overflows int64")
+		}
+		return i, bts, nil
+	default:
+		return 0, bts, msgp.TypeError{Encoded: t, Method: msgp.IntType}
+	}
+}
+
+// parseUint64Bytes parses an uint64 even if the sent value is an int64;
+// this is required because the language used for the encoding library
+// may not have unsigned types. An example is early version of Java
+// (and so JRuby interpreter) that encodes uint64 as int64:
+// http://docs.oracle.com/javase/tutorial/java/nutsandbolts/datatypes.html
+func parseUint64Bytes(bts []byte) (uint64, []byte, error) {
+	// read the generic representation type without decoding
+	t := msgp.NextType(bts)
+
+	var i int64
+	var u uint64
+	var err error
+	switch t {
+	case msgp.UintType:
+		u, bts, err = msgp.ReadUint64Bytes(bts)
+		if err != nil {
+			return 0, bts, err
+		}
+		return u, bts, err
+	case msgp.IntType:
+		i, bts, err = msgp.ReadInt64Bytes(bts)
+		if err != nil {
+			return 0, bts, err
+		}
+		return uint64(i), bts, nil
+	default:
+		return 0, bts, msgp.TypeError{Encoded: t, Method: msgp.IntType}
+	}
+}
+
+// parseInt32Bytes parses an int32 even if the sent value is an uint32;
+// this is required because the encoding library could remove bytes from the encoded
+// payload to reduce the size, if they're not needed.
+func parseInt32Bytes(bts []byte) (int32, []byte, error) {
+	// read the generic representation type without decoding
+	t := msgp.NextType(bts)
+
+	var i int32
+	var u uint32
+	var err error
+	switch t {
+	case msgp.IntType:
+		i, bts, err = msgp.ReadInt32Bytes(bts)
+		if err != nil {
+			return 0, bts, err
+		}
+		return i, bts, nil
+	case msgp.UintType:
+		u, bts, err = msgp.ReadUint32Bytes(bts)
+		if err != nil {
+			return 0, bts, err
+		}
+
+		// force-cast
+		i, ok := castInt32(u)
+		if !ok {
+			return 0, bts, errors.New("found uint32, overflows int32")
+		}
+		return i, bts, nil
+	default:
+		return 0, bts, msgp.TypeError{Encoded: t, Method: msgp.IntType}
+	}
+}