--- a/pkg/trace/pb/decoder_bytes.go
+++ b/pkg/trace/pb/decoder_bytes.go
@@ -1,3 +1,8 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-2020 Datadog, Inc.
+
 package pb
 
 import (