--- a/pkg/trace/pb/trace.go
+++ b/pkg/trace/pb/trace.go
@@ -5,6 +5,9 @@
 
 package pb
 
+//go:generate msgp -file=span.pb.go -o span_gen.go -io=false
+//go:generate msgp -io=false
+
 // Trace is a collection of spans with the same trace ID
 type Trace []*Span