… pkg/trace/pb/decoder_bytes_test.go
… pkg/trace/pb/span_gen_modifs_test.go
… pkg/trace/pb/span_gen_test.go
… pkg/trace/pb/trace_gen.go
… pkg/trace/pb/trace_gen_test.go
… pkg/trace/test/runner.go
… releasenotes/notes/trace-agent-change-msgpack-deserializer-688b3e30a4cf38c2.yaml