### What does this PR do?

This PR changes the msgpack decoding implementation to use the `msgp.Marshaler` interface rather than `msgp.Decodable`. That's a tradeoff though, since we need to allocate memory for all the bytes of the received payload at once it will increase memory usage.

An extra benefit is that the utf8 normalization will no longer allocate an intermediate string for strings which need to be mutated.

It also opens opportunities for further optimizations:
- no intermediate string allocations for truncation and normalization (directly read the strings from the raw byte buffer)
- string interning (we could intern some strings directly in the decoding layer.

Benchmarks:

```
name              old time/op    new time/op    delta
DecoderMsgpack-4    19.0ms ± 2%    12.8ms ± 1%  -32.30%  (p=0.008 n=5+5)

name              old alloc/op   new alloc/op   delta
DecoderMsgpack-4    9.12MB ± 0%    9.15MB ± 0%   +0.36%  (p=0.008 n=5+5)

name              old allocs/op  new allocs/op  delta
DecoderMsgpack-4      129k ± 0%      129k ± 0%     ~     (all equal)
```

### Motivation

What inspired you to submit this pull request?

### Additional Notes

Anything else we should know when reviewing?

### Describe your test plan

Write there any instructions and details you may have to test your PR.